Stabilize context-limit helper coverage
